univ_oahpa/requirements.txt
Django==1.4.10
Markdown==2.4
MySQL-python==1.2.4
Werkzeug==0.9.4
django-extensions==1.1.0
djangorestframework==2.3.13
## FIXME: could not find svn URL in dependency_links for this package:
flup==1.0.3.dev-20110405
ipython==1.2.1
pexpect==2.4
six==1.2.0
wsgiref==0.1.2
